requirements.txt
langchain-ollama
langchain-core
pyyaml
